update (SMAPI 3.16.2, SDV 1.5.6)

+++ source/SpriteMaster/SetVersion.rb
@@ -117,7 +117,7 @@ class SemanticVersion
 				tag = elements[0] if tag.nil?
 				tag_version = elements[1] if tag_version.nil?
 				unless tag_version.nil?
-					tag_version = tag_version.to_i
+					tag_version = tag_version.to_f
 					tag_build = tag_version
 					case tag
 						when nil
